+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-11 09:58:15
+
+### Changed Files
+- `src/features/training-modules/components/learner/AssignmentRow.tsx`
+- `src/features/training-modules/components/learner/AssignmentsTab.tsx`
+- `src/features/training-modules/components/learner/CategorySection.tsx`
+- `src/features/training-modules/components/learner/MyTrainingPage.tsx`
+
+### Statistics
+```
+ 4 files changed, 392 insertions(+), 11 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-11 09:12:23
 
 ### Changed Files